Process of Finding the Center of the LED
Changes the script so that, instead of finding the maximum of the image, it finds where the image is equivalent to the most saturated value possible (255). It then attempts to find the center of the saturated pixels and identify the row and column where this occurs. Currently there seems to be an issue with finding the row/column. Need to work on this.